+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## Version 7.2
 
+## Version 7.2.0.1 - 2024-05-15
+
+-   Added array, JSON, and vector support to UDFs
+    *(requires Kinetica v7.2.0.8 or later)*
+
+
+
 ## Version 7.2.0.0 - 2024-03-04
 
 -   Version release
+
 
 
 ## Version 7.1
@@ -28,6 +36,7 @@
 ### Version 7.1.0.0 - 2020-08-03
 
 
+
 ## Version 7.0
 
 ### Version 7.0.0.1 - 2020-04-18
@@ -41,15 +50,18 @@
 -   Version release
 
 
+
 ## Version 6.2.0 - 2018-02-11
 
 -   Version release
+
 
 
 ## Version 6.1.0 - 2017-10-10
 
 -   Added support for DATETIME data type.
 -   Performance improvements.
+
 
 
 ## Version 6.0.0 - 2017-01-24